feat: question anwser comment

+++ src/main/java/fr/istic/m2/mcq_api/dto/AnswerCommentDTO.java
@@ -4,7 +4,7 @@ import lombok.Data;
 
 @Data
 public class AnswerCommentDTO {
-    private int answerId;
+    private Long answerId;
     private String suggestion;
     private boolean isAccepted;
 }
